+++ mayan/apps/documents/tests/mixins/document_type_mixins.py
@@ -251,32 +251,6 @@ class DocumentTypeRetentionPoliciesViewTestMixin:
         )
 
 
-class DocumentTypeTestMixin:
-    auto_create_test_document_type = True
-    auto_delete_test_document_type = True
-
-    def setUp(self):
-        super().setUp()
-        self._test_document_types = []
-
-        if self.auto_create_test_document_type:
-            self._create_test_document_type()
-
-    def _create_test_document_type(self, label=None):
-        label = label or '{}_{}'.format(
-            TEST_DOCUMENT_TYPE_LABEL, len(self._test_document_types)
-        )
-
-        self._test_document_type = DocumentType.objects.create(label=label)
-        self._test_document_types.append(self._test_document_type)
-
-    def tearDown(self):
-        if self.auto_delete_test_document_type:
-            for document_type in DocumentType.objects.all():
-                document_type.delete()
-        super().tearDown()
-
-
 class DocumentTypeViewTestMixin:
     def _request_test_document_type_create_view(self):
         return self.post(